requirements.txt
pandas==2.*
datasets==2.*
renumics-spotlight==1.6.*
